# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Managment",
        "description": "Hospital Managment Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/upload-doctor": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Upload doctor information and an image",
                "description": "Uploads doctor information and an image.",
                "operationId": "uploadDoctor",
                "requestBody": {
                    "description": "Doctor details and image",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "speciality",
                                    "room",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "speciality": {
                                        "type": "string",
                                        "example": "Cardiologist"
                                    },
                                    "room": {
                                        "type": "string",
                                        "example": "Room 101"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "base64-encoded-image-data"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor Added Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/doctors/{id}": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Update doctor details",
                "description": "Updates the details of a doctor.",
                "operationId": "updateDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Doctor details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "speciality",
                                    "room",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "speciality": {
                                        "type": "string",
                                        "example": "Cardiologist"
                                    },
                                    "room": {
                                        "type": "string",
                                        "example": "Room 101"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "base64-encoded-image-data"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor details updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete a doctor",
                "description": "Deletes a doctor by ID and redirects back.",
                "operationId": "deleteDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects back after successful deletion"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Show a list of doctors",
                "description": "Displays a list of doctors.",
                "operationId": "showDoctorList",
                "responses": {
                    "200": {
                        "description": "View displaying a list of doctors"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/appointments": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Creates a new appointment based on user input.",
                "operationId": "createAppointment",
                "requestBody": {
                    "description": "Appointment details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "email",
                                    "date",
                                    "doctor",
                                    "phone",
                                    "message",
                                    "status"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Adhurim"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "adhurim.quku34@gmail.com"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "23.04.2023"
                                    },
                                    "doctor": {
                                        "type": "string",
                                        "example": "isuf"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0692935570"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "hello"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "In progress..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment Request Successful. We will contact you soon..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/my-appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get a list of user's appointments",
                "description": "Returns a list of appointments for the authenticated user.",
                "operationId": "getUserAppointments",
                "responses": {
                    "200": {
                        "description": "A list of user's appointments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointment"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancel an appointment",
                "description": "Cancels an appointment based on the provided ID.",
                "operationId": "cancelAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the appointment to cancel.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content (successful deletion)"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "speciality": {
                        "type": "string"
                    },
                    "room": {
                        "type": "string"
                    },
                    "imahe": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Appointment": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "doctor": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}